backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.13.0
django-filter==24.1
drf-spectacular==0.27.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2
pytz==2024.1
reportlab==4.0.9
Pillow==10.2.0
requests==2.31.0
python-decouple==3.8
django-extensions==3.2.3
gunicorn==21.2.0
